pyproject.toml
[tool.poetry]
name = "ztc"
version = "0.1.0"
description = "ZeroTouch Composition Engine - Bare-metal Kubernetes bootstrap"
authors = ["ZeroTouch Team"]
readme = "README.md"
packages = [{include = "ztc"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
typer = "^0.16.0"
rich = "^13.0.0"
pydantic = "^2.0.0"
jinja2 = "^3.0.0"
pyyaml = "^6.0.0"
aiohttp = "^3.9.0"
cryptography = "^42.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^1.3.0"
pyinstaller = "^6.18.0"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ztc = "ztc.cli:app"
